Create matched_cohort.sas
This is a SAS macro. It will match a small dataset to a large dataset based on any number of risk factors, controlling for a fixed proportion of individuals in each group (default is one-to-one). Doing thusly will standardize the distribution of risk factors in the referent population to the population of interest.